# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BotTalk API
  description: |
    BotTalk 是一个开源的微信推送服务（类似 Server酱），通过微信 iLink Bot 向用户发送消息推送。

    ## 认证方式

    - **SendKey**：用于推送 API，32 位十六进制字符串，支持路径传参、Query 传参和 Bearer Token 三种方式。
    - **Session Cookie**：用于管理 API，通过登录或扫码注册后自动设置，有效期 30 天。

    ## 响应格式

    推送 API 使用 `code` + `message` 风格；管理 API 使用 `success` 布尔值风格。
  version: 1.0.0
  contact:
    name: BotTalk
    url: https://bot-talk.com
servers:
  - url: https://bot-talk.com
    description: 托管服务
  - url: http://localhost:3000
    description: 本地开发 / 自部署
...
components:
  securitySchemes:
    SendKeyQuery:
      type: apiKey
      in: query
      name: key
      description: 通过 query 参数传递 SendKey
    SendKeyBearer:
      type: http
      scheme: bearer
      description: 通过 Authorization Bearer 头传递 SendKey
    SessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
      description: 登录 session cookie
  schemas:
...
    UnauthorizedError:
      type: object
      properties:
        error:
          type: string
          example: "Unauthorized. Please log in."
...
    Channel:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
        channel_type:
          type: string
          example: wechat_ilink
        wechat_openid:
          type: string
        bot_token:
          type: string
        context_token:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, active, inactive]
        is_default:
          type: integer
          enum: [0, 1]
        created_at:
          type: string
          format: date-time
...
    QRCodeData:
      type: object
      description: iLink API 返回的二维码数据（字段可能随上游变化）
      properties:
        qrcode:
          type: string
        qrcode_url:
          type: string
        expire_time:
          type: integer
...
paths:
...
  /api/channels/qrcode-public:
    post:
      tags: [Channels]
      summary: 生成公开二维码（扫码即注册）
      description: 生成扫码绑定二维码，用于新用户注册+绑定通道，无需登录。
      responses:
        '200':
          description: 二维码数据
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/QRCodeData'
        '500':
          description: 获取二维码失败
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    type: object
                    properties:
                      error:
                        type: string
  /api/channels/bind-status-public/{qrcode}:
    get:
      tags: [Channels]
      summary: 轮询公开扫码状态
      description: |
        轮询扫码状态。扫码完成后自动创建用户和默认通道，并建立登录 session。
        返回 status 为 pending 表示通道正在等待激活，用户需要在微信中向 Bot 发送一条消息完成激活。
      parameters:
        - name: qrcode
          in: path
          required: true
          schema:
            type: string
          description: 从 qrcode-public 获取的二维码字符串
      responses:
        '200':
          description: 扫码状态
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    oneOf:
                      - type: object
                        description: 等待扫码
                        properties:
                          status:
                            type: string
                            example: waiting
                      - type: object
                        description: 二维码过期
                        properties:
                          status:
                            type: string
                            example: expired
                      - type: object
                        description: 扫码完成
                        properties:
                          id:
                            type: integer
                          send_key:
                            type: string
                          status:
                            type: string
                            example: pending
                          wechat_openid:
                            type: string
                          is_new:
                            type: boolean
                            description: true=新用户, false=已有用户自动登录
        '403':
          description: Session 不匹配
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    type: object
                    properties:
                      error:
                        type: string
                        example: "Session mismatch"
  /api/channels:
    get:
      tags: [Channels]
      summary: 获取通道列表
      description: 获取当前用户的所有通道（不含已删除的 inactive 通道）。
      security:
        - SessionCookie: []
      responses:
        '200':
          description: 通道列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Channel'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
  /api/channels/qrcode:
    post:
      tags: [Channels]
      summary: 生成新通道二维码
      description: 为当前用户生成添加新通道的二维码。每个用户最多 10 个通道。
      security:
        - SessionCookie: []
      responses:
        '200':
          description: 二维码数据
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/QRCodeData'
        '400':
          description: 超过通道上限
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    type: object
                    properties:
                      error:
                        type: string
                        example: "Maximum 10 channels allowed"
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
  /api/channels/bind-status/{qrcode}:
    get:
      tags: [Channels]
      summary: 轮询新通道扫码状态
      description: |
        轮询新通道的扫码绑定状态。需要登录。
        支持新微信绑定、同一微信重绑、以及检测微信是否已被其他账户绑定。
      security:
        - SessionCookie: []
      parameters:
        - name: qrcode
          in: path
          required: true
          schema:
            type: string
          description: 从 /api/channels/qrcode 获取的二维码字符串
      responses:
        '200':
          description: 扫码状态
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    oneOf:
                      - type: object
                        description: iLink 返回的等待状态
                      - type: object
                        description: 新通道绑定成功
                        properties:
                          id:
                            type: integer
                          name:
                            type: string
                          status:
                            type: string
                            example: pending
                          wechat_openid:
                            type: string
                          is_default:
                            type: integer
                      - type: object
                        description: 同一微信重绑
                        properties:
                          id:
                            type: integer
                          name:
                            type: string
                          status:
                            type: string
                            example: pending
                          wechat_openid:
                            type: string
                          is_default:
                            type: integer
                          rebind:
                            type: boolean
                            example: true
                      - type: object
                        description: 微信已被其他账户绑定
                        properties:
                          error:
                            type: string
                            example: wechat_already_bound
                          message:
                            type: string
                            example: 该微信已被其他账户绑定
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
  /api/channels/{id}/activation:
    get:
      tags: [Channels]
      summary: 检查通道激活状态
      description: 检查通道是否已激活（用户在微信中发送了第一条消息）。
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 通道 ID
      responses:
        '200':
          description: 激活状态
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      activated:
                        type: boolean
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '404':
          description: 通道不存在
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    type: object
                    properties:
                      error:
                        type: string
                        example: "Channel not found"
  /api/channels/{id}:
    patch:
      tags: [Channels]
      summary: 更新通道
      description: 更新通道名称或设为默认通道。
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 通道 ID
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 新的通道名称
                is_default:
                  type: integer
                  enum: [0, 1]
                  description: 设为 1 则将此通道设为默认
      responses:
        '200':
          description: 更新后的通道
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Channel'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '404':
          description: 通道不存在
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    type: object
                    properties:
                      error:
                        type: string
                        example: "Channel not found"
    delete:
      tags: [Channels]
      summary: 删除通道
      description: 软删除通道（将状态设为 inactive）。
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 通道 ID
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "Channel deleted"
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '404':
          description: 通道不存在
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    type: object
                    properties:
                      error:
                        type: string
                        example: "Channel not found"
  /api/channels/{id}/check:
    post:
      tags: [Channels]
      summary: 检测通道存活状态
      description: 检测通道是否存活（通过服务端 poller 心跳检测，不调用 iLink API）。
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 通道 ID
      responses:
        '200':
          description: 存活状态
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      alive:
                        type: boolean
                      reason:
                        type: string
                        nullable: true
                        description: "不存活时的原因：no_poller, stopped, session_expired, heartbeat_timeout"
                      last_ok_seconds_ago:
                        type: integer
                        description: 距离上次成功心跳的秒数
                      status:
                        type: string
                        enum: [active, inactive]
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '404':
          description: 通道不存在
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    type: object
                    properties:
                      error:
                        type: string
                        example: "Channel not found"

  # ════════════════════════════════════════════════════════════════════
  #  Push Logs API
  # ════════════════════════════════════════════════════════════════════
...